requirements.txt
jupyter
pyyaml
nbstripout
matplotlib
numpy
scipy
flake8